Fixed build issues
