+++ Assets/Resources/NetworkBehaviourTypeContainer.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e3f957e7fd924d944bae72a2fdb3d7e7
+guid: b5efdec07be45e54c916b511303c9597
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 11400000
