+++ Assets/Resources/ServerObjects/_TestCube.prefab
@@ -25,6 +25,7 @@ GameObject:
   - component: {fileID: 114340251316627260}
   - component: {fileID: 114515986793166176}
   - component: {fileID: 114074685468992710}
+  - component: {fileID: 114798025031087092}
   m_Layer: 10
   m_Name: _TestCube
   m_TagString: Untagged
@@ -142,3 +143,14 @@ MonoBehaviour:
   inputsPerSec: 60
   movementSpeed: 30
   jumpHeight: 7
+--- !u!114 &114798025031087092
+MonoBehaviour:
+  m_ObjectHideFlags: 1
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 100100000}
+  m_GameObject: {fileID: 1216853482755444}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 38ea6535ed642ec43a7a0fee34953e2d, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier:
