+++ Assets/Scripts/Networking/NetworkBehaviour.cs
@@ -18,7 +18,7 @@ namespace MastersOfTempest.Networking
         {
             initialized = false;
             serverObject = GetComponent<ServerObject>();
-            typeId = NetworkManager.Instance.networkBehaviourManager.GetTypeIdOfNetworkBehaviour(GetType());
+            typeId = NetworkManager.Instance.networkBehaviourTypeContainer.GetTypeIdOfNetworkBehaviour(GetType());
 
             // NetworkBehaviour messages are managed by the GameClient, GameServer and their ServerObjects
             serverObject.AddNetworkBehaviourEvents(typeId, OnReceivedMessageRaw, OnNetworkBehaviourInitialized);
