+++ Assets/Scripts/Networking/NetworkBehaviourTypeContainer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 38ad8f704ba72ad4e99247aba2d93ee4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
